requirements.txt
streamlit==1.32.0
diffusers==0.26.3
torch==2.2.0
transformers==4.37.2
accelerate==0.27.2
pillow==10.2.0
huggingface_hub==0.20.3 